Moved some simple stuff in utils
