+++ include/utils.h
@@ -10,4 +10,20 @@ typedef struct TileCoords TileCoords;
 /* Function used by qsort */
 int cmpInt (const void * a, const void * b);
 
+struct Dims {
+	int w, h;
+};
+typedef struct Dims Dims;
+
+/* Just a callback function returning `void` bundled with optional generic data. */
+struct CallbackWithData {
+	void (*func)(void* whatever);
+	void* whatever;
+};
+typedef struct CallbackWithData CallbackWithData;
+
+void call_callback(CallbackWithData cb);
+
+int max(int a, int b);
+
 #endif
